+++ src/app/cart/cart.module.ts
@@ -1,21 +1,19 @@
 import { NgModule } from '@angular/core';
 
 import { SharedModule } from '../shared/shared.module';
-import { CartListComponent } from './components/cart-list/cart-list.component';
+import { CartRoutingModule } from './cart-routing.module';
 import { CartItemComponent } from './components/cart-item/cart-item.component';
 
 
 
 @NgModule({
   declarations: [
-    CartListComponent,
     CartItemComponent,
+    CartRoutingModule.components,
   ],
   imports: [
     SharedModule,
-  ],
-  exports: [
-    CartListComponent,
+    CartRoutingModule,
   ],
 })
 export class CartModule { }