+++ src/app/cart/index.ts
@@ -1,3 +1,6 @@
 export * from './components/cart-list/cart-list.component';
 export * from './components/cart-item/cart-item.component';
+
 export * from './services/cart.service';
+
+export * from './guards/is-cart-empty.guard';