+++ src/app/core/services/login.service.ts
@@ -1,0 +1,14 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class LoginService {
+  private currentRole = 'admin';
+
+  constructor() { }
+
+  getRole(): string {
+    return this.currentRole;
+  }
+}