+++ src/app/orders/components/process-order/process-order.component.ts
@@ -1,0 +1,14 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  templateUrl: './process-order.component.html',
+  styleUrls: ['./process-order.component.scss']
+})
+export class ProcessOrderComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}