+++ src/app/products/components/index.ts
@@ -1,3 +1,5 @@
 export * from './first/first.component';
 export * from './product/product.component';
 export * from './product-list/product-list.component';
+export * from './product-view/product-view.component';
+export * from './product-form/product-form.component';