+++ src/app/products/components/product/product.component.ts
@@ -1,4 +1,5 @@
 import { ChangeDetectionStrategy, Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
+import { Router } from '@angular/router';
 
 import { Product } from '../../models/products.model';
 
@@ -13,12 +14,24 @@ export class ProductComponent implements OnInit {
 
   @Output() buy: EventEmitter<number> = new EventEmitter<number>();
 
-  constructor() { }
+  constructor(
+    private router: Router,
+  ) { }
 
   ngOnInit(): void {
   }
 
-  onBuy(): void {
+  onBuy(event: Event): void {
+    event.stopPropagation();
     this.buy.emit(this.product.id);
+  }
+
+  onEdit(event: Event): void {
+    event.stopPropagation();
+    this.router.navigate([`/admin/product/edit`, this.product.id]);
+  }
+
+  isEditable(): boolean {
+    return this.router.url.includes('/admin');
   }
 }